Add workflow to release to NPM
We want to publish this package to both deno.land/x and npmjs.com

For deno, it's as simple as pushing a release tag to the repo and
boom, it's on deno.land/x. For NPM, we need to build and publish to
npmjs.com

This updates the npm build script to draw its version from the
NPM_VERSION environment variable so that it can take it dynamically.

Then, we add a workflow for new tags starting with `v`
like `v1.0.0` which then extracts out the number (in this case
`1.0.0`) from the tag and then uses it to set an output variable. This
can be used to set the NPM_VERSION variable in the build.