feat: preventing stack overflows with `k.tail()`
Previously inside a `shift()` instruction we would always return the
value of `k()`.  This caused another tick in the stack frame which
caused the stack to increase until an overflow was reached.  Previously
it only took about 1495 `shift()` calls inside a loop to cause a stack
overflow.

Now we provide the ability for the end-user to decide whether or not
they care about the return value of `k()` with the addition of
`k.tail()` which will continue the same stack frame instead of creating
a new one, preventing the likely-hood of a stack overflow.

The reason why we started investigating this inside `continuation` was
because of a previous issue in `redux-saga`:
https://github.com/redux-saga/redux-saga/issues/1592

Co-authored-by: Charles Lowell <cowboyd@frontside.com>