🐛 oneshot continuation should throw the same error consistently
It turns out that there was a bug with the `oneshot()` logic, where if
an error was raised, then the next time, the continuation would
successfully return undefined. This was not being caught because the
test case was broken (It did not check that the `boom()` continuation
failed, only that if it did, the message of the exception was the
same,

This fixes the test, and then in the event of a failure, saves the
error (much in the same way that the result is saved) and then
subsequent calls to this function will raise the exact same exception.

We could leave it as-is, because it doesn't seem to be causing any
problems, but it seems like the symmetric case of resolution and
rejection should be handled the same. If the continuation caches the
result, then it should also cache the exception.