Unify stack reduction
Currently there are two codepaths for resuming computations: `k()` and
`k.tail()` where the first hase a return value, but is recursive and
is subject to stack overflow, while `k.tail()` is not allowed to have
a return value, but has a more resilient stack. There are a number of
problems with this.

The first is the cognitive overhead on the API side. You have to
constantly remember that you need to call `k.tail()` instead of `k()`
or you might just introduce a stack overflow bug. The second is the
complexity of the implementation. In order to make the `k.tail()`
work, we needed to manage the call stack to be weirdly re-entrant with
a separate `Stack` object, and a check to see if the stack was
currently reducing (Stack.reducing). This also comes with a _memory
complexity_ penalty because we're creating a lot of intermediate
objects in the form of the stack itself, and also in the form of the
thunks that we are pushing and popping off the stack. There was also
memory overhead in creating four continuation functions (`k()`,
`k.tail()`, `reject()` and `reject.tail()`) instead of just two.

This avoids all that complexity by making the entire `evaluate()`
method into a single while loop that only takes up one native
JavaScript stack frame to execute no matter how deeply nested the
computation you want to run is. However, this does require a key
breaking change:

`k()` is a Computation, not a callback
===

the `k()` passed to `shift()` is of type `(value) => Compuation<T>`
instead of `(value) => T`. This means that in order to use it, you
have to either `yield* k()` or re-enter a reduction with evaluate. In
other words, reduction stacks must be explicitly created with
`evaluate()` in order to use them, for example, from a `setTimeot()`.

Otherwise, they can be used seamlessly if you are already inside of a
delimited continuation:

```ts
let result = evaluate(function* run() {
  let sum = 0;
  for (let i = 0; i < 100_000; i++) {
    sum += yield* shift<1>(function* incr(k) {
      return yield* k(1);
    });
  }
  return sum;
});

console.dir({ result });
```

the above will print => `{ result: 10000 }`;

The tradeoff for this breaking change is drastically reduced
complexity of implementation.